wifi: Increase kMaxStopCompleteWaitMs to 250 msec.
The current timeout value of 100msec for kMaxStopCompleteWaitMs
is not sufficient to stop the Wi-Fi HAL. QC Wi-Fi HAL also needs
to stop the event thread by sending an exit msg as part of stop WiFi HAL.
This stop operations are not completed with in the stipulated 100msec.

Hence , set this timeout value to 250msec.

Change-Id: I84c203a46793d7bf02986693fac403892c7bd48f
CRs-Fixed: 2357953